[Net] Made normal structure of port scanner module
