+++ CMakeLists.txt
@@ -10,7 +10,7 @@ set(CMAKE_CXX_STANDARD_REQUIRED True)
 add_executable(${TARGET})
 
 # target_compile_options(${TARGET} PRIVATE -fsanitize=address -g)
-# target_link_options(${TARGET} PRIVATE -fsanitize=address -g)
+target_link_options(${TARGET} PRIVATE -lpthread)
 
 enable_testing()
