+++ src/CMakeLists.txt
@@ -1,3 +1,4 @@
 target_sources(${TARGET} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/main.cpp)
 
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/Core)
+add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/Net)
