+++ src/Net/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(${TARGET} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/portScanner.cpp)
+
+target_sources(${TEST_TARGET} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/portScanner.cpp)
